WIP use modern syntax to configure omrelp
As part of poking at slow rsyslog, I found this maybe voodoo fix.

By specifying the timeout option, rsyslog is able to detect the relp
connection is bad and shutdown gracefully when asked to by its init
script.

Howvever in order to do that, we need to swap over to modern rsyslog
configuration syntax for configuring the "ship to ingestor via omrelp"
action.